+++ noodles-bam/src/io/reader.rs
@@ -71,12 +71,11 @@ use crate::Record;
 /// ### `noodles_bgzf::io::MultithreadedReader`
 ///
 /// ```
-/// # use std::{fs::File, io, num::NonZero, thread};
+/// # use std::{fs::File, io};
 /// use noodles_bam as bam;
 /// use noodles_bgzf as bgzf;
 ///
-/// let worker_count = thread::available_parallelism().unwrap_or(NonZero::<usize>::MIN);
-/// let decoder = bgzf::io::MultithreadedReader::with_worker_count(worker_count, io::empty());
+/// let decoder = bgzf::io::MultithreadedReader::new(io::empty());
 /// let _reader = bam::io::Reader::from(decoder);
 /// ```
 pub struct Reader<R> {