+++ noodles-bgzf/examples/bgzf_read_multithreaded.rs
@@ -6,8 +6,6 @@ use std::{
     env,
     fs::File,
     io::{self, BufReader, BufWriter},
-    num::NonZero,
-    thread,
 };
 
 use noodles_bgzf as bgzf;
@@ -16,14 +14,10 @@ fn main() -> io::Result<()> {
     let mut args = env::args().skip(1);
 
     let src = args.next().expect("missing src");
-    let worker_count = args
-        .next()
-        .map(|s| s.parse().expect("invalid worker_count"))
-        .unwrap_or_else(|| thread::available_parallelism().unwrap_or(NonZero::<usize>::MIN));
 
     let mut reader = File::open(src)
         .map(BufReader::new)
-        .map(|file| bgzf::io::MultithreadedReader::with_worker_count(worker_count, file))?;
+        .map(bgzf::io::MultithreadedReader::new)?;
 
     let stdout = io::stdout().lock();
     let mut writer = BufWriter::new(stdout);